+++ src/collective/collectionfilter/interfaces.py
@@ -35,7 +35,7 @@ class ICollectionFilterBaseSchema(Interface):
         pattern_options={
             'basePath': utils.target_collection_base_path,
             'recentlyUsed': True,
-            # 'selectableTypes': ['Collection'],
+            'selectableTypes': ['Collection'],
         },
     )