+++ CHANGES.rst
@@ -4,7 +4,11 @@ Changelog
 3.1 (unreleased)
 ----------------
 
-- Nothing changed yet.
+Bug fixes:
+
+- constrain ``target collection`` to a configurable registry value.
+  the default is ``['Collection', ]``
+  [petschki]
 
 
 3.0 (2019-03-25)